Removing. This introduces a massive security vulnerability.
This isn't needed, and is potentially dangerous (even on a local vagrant, if the local box isn't firewalled properly and allows connections from the outside world). Also, there are a number of files that are expected to be owned by root. This could potentially cause unexpected behavior which could be difficult to track down.